+++ Draft/Ladder/deps-debian-12.sh
@@ -1,0 +1,4 @@
+#!/bin/sh
+
+# Type a space between deps: "item next next etc"
+DEPS="gnugo"